update by toankx

+++ app/Http/Controllers/PostsController.php
@@ -5,9 +5,21 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Post;
 use DB;
+use DateTime;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -42,12 +54,30 @@ class PostsController extends Controller
     {
       $this->validate( $request, [
         'title' => 'required',
-        'description' => 'required'
+        'description' => 'required',
+        'cover_image' => 'image|nullable|max:1999'
       ]);
 
+      // Kiểm tra cover_image trước khi thêm vào Database
+      if( $request->hasFile('cover_image') ) {
+        // Lấy filename với Extention
+        $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+        // Get fileName
+        $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+        // Get extention of file
+        $extention = $request->file('cover_image')->getClientOriginalExtension();
+        // filename
+        $fileNameToStore = $fileName . '_' . time() . '.' . $extention;
+        // upload image
+        $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+      }else {
+        $fileNameToStore = 'noimage.jpeg';
+      }
       $post = new Post;
       $post->title = $request->input('title');
       $post->body = $request->input('description');
+      $post->user_id = auth()->user()->id;
+      $post->cover_image = $fileNameToStore;
 
       $post->save();
       return redirect('/posts')->with('success', 'Tạo thành công bài viết.');
@@ -75,6 +105,10 @@ class PostsController extends Controller
     public function edit($id)
     {
       $post = Post::find($id);
+      if(auth()->user()->id !== $post->user_id) {
+        // return view('posts.edit')->with('error', 'Bạn không có quyền truy cập tình năng này.');
+        return redirect('/posts')->with('error', 'Bạn không có quyền truy cập tình năng này.');
+      }
       return view('posts.edit')->with('post', $post);
     }
 
@@ -93,6 +127,12 @@ class PostsController extends Controller
       ]);
 
       $post = Post::find($id);
+
+      if(auth()->user()->id !== $post->user_id) {
+        // return view('posts.edit')->with('error', 'Bạn không có quyền truy cập tình năng này.');
+        return redirect('/posts')->with('error', 'Bạn không có quyền truy cập tình năng này.');
+      }
+
       $post->title = $request->input('title');
       $post->body = $request->input('description');
 
@@ -109,6 +149,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
       $post = Post::find($id);
+      if(auth()->user()->id !== $post->user_id) {
+        return redirect('/posts')->with('error', 'Bạn không có quyền truy cập tình năng này.');
+      }
       $post->delete();
       return redirect('/posts')->with('success', 'Bài viết đã được xóa.');
     }
